move /sidekick to /
